+++ packages/@leight/mantine/src/component/index.ts
@@ -1,3 +1,4 @@
+export * from "./App";
 export * from "./DropZone";
 export * from "./Paper";
 export * from "./TabLinkMenu";